refactor(source): move pure source-identity helpers to utils/sources
sourceKey (+ djb2), sourceIdentity, and sameSourceIdentity are pure
functions on a source's (src, branch) identity with no signal dependency,
so they belong beside their siblings identityBranch/resolveBranch in
utils/sources, not in the stateful store. state/stores/source keeps the
signals, the recents mutations, and the URL effect built on them.

RecentsList now imports the two identity helpers from utils instead of
reaching into the store for pure logic; excludes + useManifestSource import
sourceKey from utils. Unit tests for the moved functions live in
tests/utils/sources.test.ts (mirroring the new home), with focused
sourceIdentity/sameSourceIdentity coverage added.
