+++ app/tests/state/stores/excludes.test.ts
@@ -1,5 +1,6 @@
 import { describe, it, expect, beforeEach } from 'vitest';
-import { CURRENT_SOURCE, sourceKey } from '@/state/stores/source';
+import { CURRENT_SOURCE } from '@/state/stores/source';
+import { sourceKey } from '@/utils/sources';
 import {
   EXCLUDES,
   ACTIVE_EXCLUDES,
